reorganized the testers for exporting applicators. honestly, who ever needs them can finish testing them
